+++ uni/lib/view/Widgets/exam_card.dart
@@ -1,11 +1,10 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_redux/flutter_redux.dart';
+import 'package:tuple/tuple.dart';
 import 'package:uni/controller/exam.dart';
 import 'package:uni/model/app_state.dart';
 import 'package:uni/model/entities/exam.dart';
-import 'package:flutter/cupertino.dart';
-import 'package:tuple/tuple.dart';
-import 'package:flutter/material.dart';
-import 'package:flutter_redux/flutter_redux.dart';
-import 'package:uni/utils/constants.dart' as Constants;
+import 'package:uni/utils/constants.dart' as constants;
 import 'package:uni/view/Widgets/date_rectangle.dart';
 import 'package:uni/view/Widgets/request_dependent_widget_builder.dart';
 import 'package:uni/view/Widgets/row_container.dart';
@@ -18,7 +17,8 @@ import 'generic_card.dart';
 class ExamCard extends GenericCard {
   ExamCard({Key? key}) : super(key: key);
 
-  ExamCard.fromEditingInformation(Key key, bool editingMode, Function onDelete)
+  const ExamCard.fromEditingInformation(
+      Key key, bool editingMode, Function()? onDelete)
       : super.fromEditingInformation(key, editingMode, onDelete);
 
   @override
@@ -26,7 +26,7 @@ class ExamCard extends GenericCard {
 
   @override
   onClick(BuildContext context) =>
-      Navigator.pushNamed(context, '/' + Constants.navExams);
+      Navigator.pushNamed(context, '/${constants.navExams}');
 
   /// Returns a widget with all the exams card content.
   ///
@@ -50,7 +50,7 @@ class ExamCard extends GenericCard {
         status: examsInfo.item2,
         contentGenerator: generateExams,
         content: examsInfo.item1,
-        contentChecker: examsInfo.item1 != null && examsInfo.item1.isNotEmpty,
+        contentChecker: examsInfo.item1.isNotEmpty,
         onNullContent: Center(
           child: Text('Não existem exames para apresentar',
               style: Theme.of(context).textTheme.headline6),
@@ -63,7 +63,7 @@ class ExamCard extends GenericCard {
   Widget generateExams(exams, context) {
     return Column(
       mainAxisSize: MainAxisSize.min,
-      children: this.getExamRows(context, exams),
+      children: getExamRows(context, exams),
     );
   }
 
@@ -72,11 +72,12 @@ class ExamCard extends GenericCard {
   List<Widget> getExamRows(context, exams) {
     final List<Widget> rows = <Widget>[];
     for (int i = 0; i < 1 && i < exams.length; i++) {
-      rows.add(this.createRowFromExam(context, exams[i]));
+      rows.add(createRowFromExam(context, exams[i]));
     }
     if (exams.length > 1) {
       rows.add(Container(
-        margin: EdgeInsets.only(right: 80.0, left: 80.0, top: 15, bottom: 7),
+        margin:
+            const EdgeInsets.only(right: 80.0, left: 80.0, top: 15, bottom: 7),
         decoration: BoxDecoration(
             border: Border(
                 bottom: BorderSide(
@@ -84,7 +85,7 @@ class ExamCard extends GenericCard {
       ));
     }
     for (int i = 1; i < 4 && i < exams.length; i++) {
-      rows.add(this.createSecondaryRowFromExam(context, exams[i]));
+      rows.add(createSecondaryRowFromExam(context, exams[i]));
     }
     return rows;
   }
@@ -93,21 +94,19 @@ class ExamCard extends GenericCard {
   /// others in the card).
   Widget createRowFromExam(context, Exam exam) {
     return Column(children: [
-      DateRectangle(date: exam.weekDay + ', ' + exam.day + ' de ' + exam.month),
-      Container(
-        child: RowContainer(
-          color: isHighlighted(exam)
-              ? Theme.of(context).backgroundColor
-              : Theme.of(context).hintColor,
-          child: ScheduleRow(
-            subject: exam.subject,
-            rooms: exam.rooms,
-            begin: exam.begin,
-            end: exam.end,
-            type: exam.examType,
-            date: exam.date,
-            teacher: '',
-          ),
+      DateRectangle(date: '${exam.weekDay}, ${exam.day} de ${exam.month}'),
+      RowContainer(
+        color: isHighlighted(exam)
+            ? Theme.of(context).backgroundColor
+            : Theme.of(context).hintColor,
+        child: ScheduleRow(
+          subject: exam.subject,
+          rooms: exam.rooms,
+          begin: exam.begin,
+          end: exam.end,
+          type: exam.examType,
+          date: exam.date,
+          teacher: '',
         ),
       ),
     ]);
@@ -117,13 +116,13 @@ class ExamCard extends GenericCard {
   /// date exam with a separator between them.
   Widget createSecondaryRowFromExam(context, exam) {
     return Container(
-      margin: EdgeInsets.only(top: 8),
+      margin: const EdgeInsets.only(top: 8),
       child: RowContainer(
         color: isHighlighted(exam)
             ? Theme.of(context).backgroundColor
             : Theme.of(context).hintColor,
         child: Container(
-          padding: EdgeInsets.all(11),
+          padding: const EdgeInsets.all(11),
           child: Row(
               mainAxisAlignment: MainAxisAlignment.spaceBetween,
               mainAxisSize: MainAxisSize.max,