+++ uni/lib/view/Widgets/other_links_card.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:uni/view/Widgets/link_button.dart';
+
 import 'generic_card.dart';
 
 /// Manages the 'Current account' section inside the user's page (accessible
@@ -7,13 +8,13 @@ import 'generic_card.dart';
 class OtherLinksCard extends GenericCard {
   OtherLinksCard({Key? key}) : super(key: key);
 
-  OtherLinksCard.fromEditingInformation(
-      Key key, bool editingMode, Function onDelete)
+  const OtherLinksCard.fromEditingInformation(
+      Key key, bool editingMode, Function()? onDelete)
       : super.fromEditingInformation(key, editingMode, onDelete);
 
   @override
   Widget buildCardContent(BuildContext context) {
-    return Column(children: [
+    return Column(children: const [
       LinkButton(title: 'Impressão', link: 'https://print.up.pt')
     ]);
   }