+++ uni/lib/view/Widgets/sigarra_links_card.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:uni/view/Widgets/link_button.dart';
+
 import 'generic_card.dart';
 
 /// Manages the 'Current account' section inside the user's page (accessible
@@ -7,8 +8,8 @@ import 'generic_card.dart';
 class SigarraLinksCard extends GenericCard {
   SigarraLinksCard({Key? key}) : super(key: key);
 
-  SigarraLinksCard.fromEditingInformation(
-      Key key, bool editingMode, Function onDelete)
+  const SigarraLinksCard.fromEditingInformation(
+      Key key, bool editingMode, Function()? onDelete)
       : super.fromEditingInformation(key, editingMode, onDelete);
 
   @override