+++ uni/linux/flutter/generated_plugins.cmake
@@ -4,7 +4,6 @@
 
 list(APPEND FLUTTER_PLUGIN_LIST
   sentry_flutter
-  sqlite3_flutter_libs
   url_launcher_linux
 )