+++ uni/windows/flutter/generated_plugins.cmake
@@ -5,7 +5,6 @@
 list(APPEND FLUTTER_PLUGIN_LIST
   connectivity_plus_windows
   sentry_flutter
-  sqlite3_flutter_libs
   url_launcher_windows
 )